+++ RELEASE_NOTES.md
@@ -1,4 +1,5 @@
 - Add example to run a local python script in the pdal_ign_plugin docker image
+- Add parameters and "main" to function "add points (.GeoJSON) in pointcloud (.LAZ / .LAS)"
 
 ### 0.3.1